# Restore SDPA in Gemma2 models for transformers > 4.45