--- a/tests/openvino/test_modeling.py
+++ b/tests/openvino/test_modeling.py
@@ -1101,7 +1101,7 @@ class OVModelForCausalLMIntegrationTest(unittest.TestCase):
 
         # starting from transformers 4.45.0 gemma2 uses eager attention by default, while ov - sdpa
         if model_arch == "gemma2" and is_transformers_version(">=", "4.45.0"):
-            model_kwargs["attn_implemenation"] = "sdpa"
+            model_kwargs["attn_implementation"] = "sdpa"
         # Qwen tokenizer does not support padding, chatglm, glm4 testing models produce nan that incompatible with beam search
         if model_arch in ["qwen", "chatglm", "glm4"]:
             return