# What does this PR do?
transformers 4.45 introduces perf regression for gemma2 models due to switching default realization from sdpa to eager in this commit:
https://github.com/huggingface/transformers/commit/975b988bfe6e7ebb47390cd9a1556c6888804883

benchmarking results for CPU:

| model |  precision | input prompt len | 2sd token latency (ms) without sdpa | 2nd token latency (ms) with sdpa |
----------|---------------|-----------------------|-------------------------------------------------|---------------------------------------|
|  gemma-2-2b  | FP16 | 32 | 37.9 | 26.7|
| gemma-2-2b | FP16 | 1024 | 106.7 | 27.3|
| gemma-2-9b | FP16 | 32 | 112.4 | 82.3 |
| gemma-2-9b | FP16 | 1024 | 310.2 | 83.7 |
|  gemma-2-2b  | INT8 | 32 | 31.9 | 20.9|
| gemma-2-2b | INT8 | 1024 | 111.3 | 21.3|
| gemma-2-9b | INT8 | 32 | 82.9 | 53.4 |
| gemma-2-9b | INT8 | 1024 | 301.9 | 55.5 |

## Before submitting
- [ ] This PR fixes a typo or improves the docs (you can dismiss the other checks if that's the case).
- [ ] Did you make sure to update the documentation with your changes?
- [ ] Did you write any new necessary tests?