@IlyasMoutawwakil 2024-10-28
what about the issue https://github.com/huggingface/transformers/issues/32848 which is the reason behind the  default implementation change. Can you make sure the snippet in the issue returns the correct outputs with sdpa+ov_model ?