@eaidova 2024-10-28
@IlyasMoutawwakil , I tested gemma-2-2b-it, output seems correct:

['Let\'s count them! \n\nYou\'ve used the word "laugh" **one** times in your question. 😊 \n', 'There are **two** "laugh" words in your sentence. 😊 \n', 'There are **three** "laugh" words in the phrase "laugh laugh laugh". \n', 'There are **four** "laugh" words in the phrase "laugh laugh laugh laugh". \n', 'There are **5** "laugh" words in the phrase "laugh laugh laugh laugh laugh". \n']